Allow layouts to specify properties to save
In anticipation of saving sort order and column separator position
