+++ app/src/main/java/edu/wpi/first/shuffleboard/app/json/WidgetSaver.java
@@ -16,7 +16,6 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.google.gson.JsonSerializationContext;
 
-import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Arrays;
@@ -25,7 +24,6 @@ import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import javafx.beans.property.Property;
 
@@ -35,6 +33,7 @@ public class WidgetSaver implements ElementTypeAdapter<Widget> {
   private static final Logger log = Logger.getLogger(WidgetSaver.class.getName());
 
   private final SourcedRestorer sourcedRestorer = new SourcedRestorer();
+  private final PropertySaver propertySaver = new PropertySaver();
 
   @Override
   public JsonElement serialize(Widget src, JsonSerializationContext context) {
@@ -45,46 +44,22 @@ public class WidgetSaver implements ElementTypeAdapter<Widget> {
     }
     object.addProperty("_title", src.getTitle());
 
-    final List<Property<?>> savedProperties = getPropertyFields(src)
+    final List<Property<?>> savedProperties = PropertySaver.getPropertyFields(src.getClass())
         .map(f -> ReflectionUtils.<Property<?>>getUnchecked(src, f))
         .collect(Collectors.toList());
 
     // Save exported properties
     for (Property p : src.getProperties()) {
       if (!savedProperties.contains(p)) {
-        serializeProperty(context, object, p, p.getName());
+        PropertySaver.serializeProperty(context, object, p, p.getName());
       }
     }
 
     // Save @SaveThisProperty fields
-    getPropertyFields(src)
-        .forEach(f -> {
-          Property<?> property = ReflectionUtils.getUnchecked(src, f);
-          String name = getSavedName(property, f.getAnnotation(SaveThisProperty.class));
-          serializeProperty(context, object, property, name);
-        });
+    propertySaver.saveAnnotatedFields(src, context, object);
 
     // Save @SavePropertyFrom fields
-    Arrays.stream(src.getClass().getDeclaredFields())
-        .filter(f -> f.getAnnotationsByType(SavePropertyFrom.class).length > 0)
-        .forEach(f -> {
-          SavePropertyFrom[] annotations = f.getAnnotationsByType(SavePropertyFrom.class);
-          for (SavePropertyFrom annotation : annotations) {
-            if (annotation.propertyName().isEmpty()) {
-              throw new IllegalArgumentException("No property name was specified");
-            }
-            String savedName = getPropertyName(annotation);
-            String propName = annotation.propertyName();
-            try {
-              Object field = ReflectionUtils.getUnchecked(src, f);
-              Method getter = getGetter(f.getType(), propName);
-              Object value = getter.invoke(field);
-              object.add(savedName, context.serialize(value, getter.getReturnType()));
-            } catch (ReflectiveOperationException e) {
-              throw new RuntimeException("Could not get value of property '" + propName + "' of " + f.getType(), e);
-            }
-          }
-        });
+    propertySaver.saveNestedProperties(src, context, object);
 
     return object;
   }
@@ -101,7 +76,7 @@ public class WidgetSaver implements ElementTypeAdapter<Widget> {
       widget.setTitle(title.getAsString());
     }
 
-    List<Property<?>> savedProperties = getPropertyFields(widget)
+    List<Property<?>> savedProperties = PropertySaver.getPropertyFields(widget.getClass())
         .map(f -> ReflectionUtils.<Property<?>>getUnchecked(widget, f))
         .collect(Collectors.toList());
 
@@ -117,43 +92,10 @@ public class WidgetSaver implements ElementTypeAdapter<Widget> {
     }
 
     // Load @SaveThisProperty fields
-    getPropertyFields(widget)
-        .forEach(f -> {
-          Property property = ReflectionUtils.getUnchecked(widget, f);
-          String name = getSavedName(property, f.getAnnotation(SaveThisProperty.class));
-          Object deserialize = context.deserialize(obj.get(name), property.getValue().getClass());
-          // Can be null if not set
-          // Instead of throwing IllegalArgumentException, just let the default value be used
-          if (deserialize != null) {
-            property.setValue(deserialize);
-          }
-        });
+    propertySaver.readAnnotatedFields(widget, context, obj);
 
     // Load @SavePropertyFrom fields
-    Arrays.stream(widget.getClass().getDeclaredFields())
-        .filter(f -> f.getAnnotationsByType(SavePropertyFrom.class).length > 0)
-        .forEach(f -> {
-          SavePropertyFrom[] annotations = f.getAnnotationsByType(SavePropertyFrom.class);
-          for (SavePropertyFrom annotation : annotations) {
-            if (annotation.propertyName().isEmpty()) {
-              throw new IllegalArgumentException("No property name set");
-            }
-            String savedName = getPropertyName(annotation);
-            String propName = annotation.propertyName();
-            try {
-              Object field = ReflectionUtils.getUnchecked(widget, f);
-              Method set = getSetter(f.getType(), propName);
-              Object deserialize = context.deserialize(obj.get(savedName), set.getParameterTypes()[0]);
-              // Can be null if not set
-              // Instead of throwing IllegalArgumentException, just let the default value be used
-              if (deserialize != null) {
-                set.invoke(field, deserialize);
-              }
-            } catch (IllegalAccessException | InvocationTargetException e) {
-              throw new RuntimeException("Could not get value of property '" + propName + "' of " + f.getType(), e);
-            }
-          }
-        });
+    propertySaver.readNestedProperties(widget, context, obj);
 
     // Load sources last - widgets can have saved properties that modify sources when they're added
     for (int i = 0; i > Integer.MIN_VALUE; i++) {
@@ -185,127 +127,6 @@ public class WidgetSaver implements ElementTypeAdapter<Widget> {
     }
 
     return widget;
-  }
-
-  /**
-   * Gets the getter method in the given class for a property with the given name, in the form {@code get<name>} e.g.
-   * {@code getFoo()}. The method must be public. Boolean properties can also be in the form {@code is<name>}, e.g.
-   * {@code isFoo()}.
-   *
-   * @param clazz        the class containing the property
-   * @param propertyName the name of the property
-   *
-   * @return the getter method for the property
-   *
-   * @throws IllegalArgumentException if there is not exactly ONE method matching {@code public Foo get<property>()}
-   *                                  or {@code public boolean is<property>()}
-   */
-  static Method getGetter(Class<?> clazz, String propertyName) {
-    // Convert the first character to uppercase, so property name "foo" becomes "getFoo" or "isFoo"
-    String base = Character.toUpperCase(propertyName.charAt(0)) + propertyName.substring(1);
-    String getterName = "get" + base;
-    String isName = "is" + base;
-    List<Method> possibleGetters = Arrays.stream(clazz.getMethods())
-        .filter(m -> m.getName().equals(getterName) || isBooleanGetterWithName(m, isName))
-        .filter(m -> m.getParameterCount() == 0)
-        .collect(Collectors.toList());
-    if (possibleGetters.isEmpty()) {
-      throw new IllegalArgumentException("No getters for property '" + propertyName + "' in " + clazz);
-    } else if (possibleGetters.size() > 1) {
-      throw new IllegalArgumentException("Multiple getters for property " + propertyName + ": " + possibleGetters);
-    }
-    return possibleGetters.get(0);
-  }
-
-  private static boolean isBooleanGetterWithName(Method method, String mName) {
-    return (method.getReturnType() == boolean.class || method.getReturnType() == Boolean.class)
-        && method.getName().equals(mName);
-  }
-
-  /**
-   * Gets the setter method in the given class for a property with the given name, in the form {@code set<name>} e.g.
-   * {@code setFoo()}. The method must be public, and cannot have any overrides (for example
-   * {@code public void setFoo(Foo)} and {@code public void setFoo(Bar)} would cause problems).
-   *
-   * @param clazz        the class containing the property
-   * @param propertyName the name of the property
-   *
-   * @return the setter method for the property
-   *
-   * @throws IllegalArgumentException if there is not exactly ONE method matching
-   *                                  {@code public void set<property>(<type>)}
-   */
-  static Method getSetter(Class<?> clazz, String propertyName) {
-    // Convert the first character to uppercase, so property name "foo" becomes "setFoo"
-    String setterName = "set" + Character.toUpperCase(propertyName.charAt(0)) + propertyName.substring(1);
-    List<Method> possibleSetters = Arrays.stream(clazz.getMethods())
-        .filter(m -> m.getName().equals(setterName))
-        .filter(m -> m.getParameterCount() == 1)
-        .filter(m -> m.getReturnType() == void.class)
-        .collect(Collectors.toList());
-    if (possibleSetters.isEmpty()) {
-      throw new IllegalArgumentException("No setter for property " + propertyName + " in " + clazz.getName());
-    } else if (possibleSetters.size() > 1) {
-      throw new IllegalArgumentException("Too many setters for property " + propertyName + ": "
-          + possibleSetters.stream()
-          .map(Method::getName)
-          .collect(Collectors.joining(", ")));
-    }
-    return possibleSetters.get(0);
-  }
-
-  /**
-   * Gets a stream of the fields in the widget's class annotated with {@link SaveThisProperty @SaveThisProperty} that
-   * are not available in its {@link Widget#getProperties() properties list}.
-   */
-  static Stream<Field> getPropertyFields(Widget widget) {
-    return Arrays.stream(widget.getClass().getDeclaredFields())
-        .filter(f -> f.isAnnotationPresent(SaveThisProperty.class))
-        .filter(f -> Property.class.isAssignableFrom(f.getType()))
-        .peek(f -> f.setAccessible(true));
-  }
-
-  static void serializeProperty(JsonSerializationContext context, JsonObject object, Property p, String name) {
-    object.add(name, context.serialize(p.getValue()));
-  }
-
-  /**
-   * Gets the saved name a property annotated with {@code @SaveThisProperty}. If the name specified in the annotation
-   * is an empty string (the default value), then the property's name will be used. Otherwise, the saved name will be
-   * the one specified in the annotation.
-   *
-   * @param property   the property to get the saved name of
-   * @param annotation the annotation on the property's field
-   *
-   * @return the saved name of the property
-   *
-   * @throws IllegalArgumentException if the property has no name (either null or empty string) and the annotation does
-   *                                  not specify the property name
-   */
-  static String getSavedName(Property<?> property, SaveThisProperty annotation) {
-    if (annotation.name().isEmpty()) {
-      if (property.getName() == null || property.getName().isEmpty()) {
-        throw new IllegalArgumentException("The property has no name, and no name was specified in the annotation");
-      }
-      return property.getName();
-    } else {
-      return annotation.name();
-    }
-  }
-
-  /**
-   * Gets the saved name of a property saved from an annotated property field.
-   *
-   * @param annotation the annotation to get the name of the property from
-   *
-   * @return the saved name of the property as set by the annotation
-   */
-  static String getPropertyName(SavePropertyFrom annotation) {
-    if (annotation.savedName().isEmpty()) {
-      return annotation.propertyName(); // Don't need to check this - it's done before this method is called
-    } else {
-      return annotation.savedName();
-    }
   }
 
 }
